Set a persistent default home page for the browser.
Fix for buganizer issue 1793708: 'Set home page' is empty after resetting to default.  Keep track of the base for the default home page in a final variable.  In loadFromDb, add the client to that (instead of to homeUrl, which could result in the client showing up twice in homeUrl) to get the homeUrl.  Then, in resetDefaultPreferences, call setHomePage, with that same construction, so if the user clears their settings, they will return to having the default home page.