Fix Google search suggestions in Browser address bar.
Before, the BrowserProvider talked directly to the
SearchManagerService and implemented its own version of
the SearchManager code for getting search suggestions.
This breaks if the global search provider is not
GoogleSearch.

This patch changes BrowserProvider to instead use
the SearchManager API to get suggestions from
the GoogleSearch application. When we have support
for third-party web search providers, this should be
changed to use the selected web search provider.