When adding a bookmark, use the same database entry as the history item.
Fix for buganizer issue 1659694.  Prior to this change, the user could
visit http://cuil.com, and then bookmark http://www.cuil.com, and the
bookmark would be stored in the database as a separate entry.  This
results in the bookmark claiming to have been visited long ago.  Now,
the two are stored as one entry, clearing up disagreements about
number of visits and last visited date.