Swap the title and subtitle of browser suggestions, so the webpage title shows as the first line and the url shows as the second line. Preceding "http://" is stripped from urls. If the first line would be empty or whitespace, the second line is promoted to the first line.
This affects suggestions both for the browser and for global search. I
thought this looked a lot better so ran it by Leland and he's cool with
this.

I went ahead and reused the raw int values being used for column indices
and such, but this is bad. You guys should really consider making this
code a little safer/cleaner. :)