Do not display unvisited sites in the history.
Fix for issue 1831854: An unvisited, bookmarked site
shows up in the history under "1 month ago".  This is
because although the site will be in the database with
a visited date of 0, it has some visits, so that it
will raise its priority in other lists.  Now, ignore
any sites with a visited date of 0 (or less) in the
history page.  Make the HistoryAdapter's member
variables private, and make the where clause final.

	modified:   src/com/android/browser/BrowserHistoryPage.java
	modified:   src/com/android/browser/BrowserHistoryPage.java