Do not use the original url in onReceivedTitle.
We used to use the original url because that was what was entered into the
history database. Now, with a change in webkit, the correct url will be passed
to updateVisitedHistory if the server sent a redirect. This allows for the
correct url and title to be displayed in the title bar as well as the correct
url and title to be stored in the history.