Updates to the browser's custom title bar.
Use an asset copied from the Market for the background.  Since this background is
dark, change the title's text to white, as well as the dividing line.  Changed
the functionality of the buttons on the title bar.  Also flipped the title and
the url, and fixed a bug where "Loading..." remained for too long.  Also added
the drop shadow below the title bar.