Update browser to use new Intent URI expansion.
This changes the browser's URI->intent processing to use the new generic
"intent:" scheme supported by Intent.  Doing so allows the user to
provide links to arbitrary intents in a web page.  The browser restricts
which intents can actually be execute to those supported by the
BROWSABLE category by adding this to the resulting Intent and making
sure there is no explicit component in the Intent.

With the addition of package-specific Intents, this allows people to
have a link that is guaranteed to launch an activity in their own
package, not allowing others to intercept it.