Pass through SearchManager.APP_DATA bundle in web searches.
This got missed when we made the change to invoke the system default web search handler instead of the browser directing all searches to Google. This makes sure that the receiving web search providers can pass on the source parameter appropriately as part of the results page that they launch.