Skeleton backup agent for Browser
This is a skeleton backup agent for the Android Browser, intended to back up and
restore the bookmarks table from browser.db by extracting the bookmarks table
into a flattened file representation on backup, then unflattening and rebuilding
the table during restore.

It compiles but is completely untested, since the db code to read/write the
bookmarks table is not finished.  The backup agent has been declared in the
Browser manifest, however.