Show the new tab when creating it from the bookmarks/history.
When we are in the tab picker and a new tab is created from bookmarks/history,
show the new tab in the picker and then animate to it. Only do this for the
"open tabs in background" setting since openTab() will show the tab if that
setting is false.

We animate to the new tab because we currently don't have a way to show the new
tab in the tab picker with updated picker data (title and url in particular).
So instead of showing a blank tab, animate to the tab. Once the animations have
been redesigned and the code refactored, this will hopefully get better.