Integrate bookmark grid view mode with list view mode.
This way the two different modes can share code and have
all the same behaviors (i.e. context menu options).

AndroidManifest.xml
Remove BookmarkGridPage, which is no longer an Activity.

bookmarks.xml
Add a menu option to switch between the different viewing modes.

strings.xml
Name for that menu option.

AddNewBookmark.java
Remove copyTo, which is not necessary.

BookmarkGridPage.java
Change to just be a GridView, with our desired settings, and
overriding onSizeChanged to size the thumbs appropriately.  Now its
constructor takes a BrowserBookmarksAdapter, which it uses instead
of its old private Adapter.

BrowserBookmarksAdapter.java
Now has two modes, so it can be used for both the bookmarks grid
and the list view.  Removed a bunch of obsolete code, which was
used for searching/sorting the bookmarks.

BrowserBookmarksPage.java
Now switches back and forth between using a GridView and a ListView.