Work on the Bookmark Grid, including the favicon.
Remove the file BookmarkGridPage, since a GridView will do just fine.
Now we specify a height and width for the bitmaps rather than adjusting
them for the screen size.  Also add the favicon and match the design
more closely for the Add option.  In BrowserActivity, match the size
of the saved screenshots to the desired size for the Bookmarks
Activity.  Also move the logic into its own method so it can be
called from multiple places (still called in only one at the moment).