Updates for the look of the bookmarks page.
Add new assets.

-tab_indicator.xml
State selector for the background of tabs in the CombinedHistoryActivity

-bookmark_thumnail.xml
Web page title is now above the thumb.  Removed the favicon.  Added a
shadow.

-tab_header.xml
Layout for the tabs in CombinedHistoryActivity

-tabs.xml
Change the height of the tabs.

-bookmarks.xml
Remove the title for the "switch" menu item, which gets set in Java,
depending on the current viewing mode.

-strings.xml
Remove the placeholder menu item text, add the actual text for
switching between modes.

-BrowserBookmarksAdapter
No longer set the favicon on top of the bookmark thumbs in grid mode.

-BrowserBookmarksPage
Implement onPrepareOptionsMenu, and set the title of a menu option,
depending on the current viewing state.

-CombinedHistoryActivity
Create a common function for setting up all the tabs.  Change the
look of the tabs to match the new design.