Make the "Most visited" page have a thumbnail viewing mode.
Start capturing screenshots for all visited pages, and show them
in a grid view on the most visited page.  Modify the BrowserBookmarks-
Page and Adapter so they can be used for the "Most visited" page.
Also change some of the ids for context menu items to be the same
in the browser and history context menus, for simplicity in
handling them in BrowserBookmarksPage.