Do not include ourselves as a candidate for downloading content.
If we are attempting to download a file and try to resolve to another activity,
it is possible to get BrowserActivity as the best candidate. In that case, we
need to fall back to the download manager. If we do not, we end up in a loop
trying to download the file and sending ourselves a new intent.

BUG: 2062310