Show the title bar when pressing bringing up the menu.
Show an actionable title bar at the top of the screen when the
menu shows.

Change TitleBar* to take a Bitmap for setFavicon, and move the
logic for creating the drawable into TitleBar, so that we can
update directly from the WebView (whose getFavicon method returns
a bitmap).

Remove CUSTOM_BROWSER_BAR, since we are set on using our custom
title bar, and it enabled me to put mGenericFavicon in TitleBar.

Also take screen density into account when sizing the icons for
the title bar.

Change-Id: I3ee6b1d02e8845299f9cd1f648b271543683954a