Do not show the title bar if the Activity is paused.
Fix http://b/issue?id=2121363

We attempt to show the fake title bar as a result of
a callback.  It is possible that we receive this callback
while the BrowserActivity is not running (not sure how
to reproduce, exactly, but the monkey did it; see the
bug).  Before showing the title bar, do a check to see
if the BrowserActivity is currently paused, and only
show it if not.

Change-Id: I770a1f86ae50f936f45bbf1dc2f047049b8e1aa3