Display favicons for history items.
Use the data from the java database before talking to the native database. The
bookmarks database may contain favicons that the native database does not have a
mapping for. Used the native icon database for recent history items since those
are not stored in the java db. Request the icons each time the activity is
created since the url->bitmap mapping is a global.