Account for find dialog height in content height
Use the set dialog height when the find dialog is up so that
found items at the bottom of the screen are scrolled into view.

This uses a companion change in webview that takes the find
dialog height into account when computing the content height.

Fixes http://b/issue?id=2122916