Add a shadow to the fake title bar.
Remove old code which requests windowOverlays.  When inserting the
fake title bar, put a shadow underneath it.  Do not show it if the
scroll is at the top, since it is transparent, and showing it along
with the shadow below the real title bar would show a darker shadow.

Fixes http://b/issue?id=2118813

Change-Id: I84b8975ea82cc8659b702112c7e67498f863916c