Update the fake title bar appropriately.
Fix for http://b/issue?id=2168682

Rather than updating the fake title bar when it is shown,
always update it with the regular title bar, so it will
have the correct icon etc when it is shown.  This also
means that we keep it around, which has the bonus of
eliminating several creations and gcs.