Move Tab out of TabControl to its own class. So that each tab has its own WebViewClient, WebChromeClient, ErrorConsoleView, lock icon type and inLoad (matching mPageStarted) state.
Clean up BrowserActivity, TabControl, create a new Tab class.

Fix stopAllLoading() to stop all main window and sub window loading.

Remove mScale/mScrollX/Y from PickerData as it is not used any more.

Remove doFlick/SensorManager as it is not used any more.

Remove whiteList as it is not used any more.

Remove MASF proxy service as it is not used any more.

With this change,
we will take the snapshot even when it is in the background.
we will update the touchicon url even when it is in the background.
we will update the bookmark/history even when it is in the background.
we will update database quota and max appCache size even when it is in the background.
we will show the error dialog of the background tab when it is brought to the foreground. (New behavior)
we will update the lock icon correctly when switching tabs.
we will proceed the http auth request for the background tab if we already have username and password.

Fix http://b/issue?id=2158621