Only show manually entered URLs in suggestions.
In the bookmarks database, add a column to keep track of whether
the url was entered by the user.  When providing a Cursor for
suggestions, do not show visited websites that were not entered
by the user.  In BrowserActivity, add manually entered urls to
the database, so that they can be seen in the suggestons.

Requires a change to frameworks/base.

Fixes http://b/issue?id=2281371