Fix a monkey crash by pretending to be a real adapter.
The monkey could generate this crash by sending a touch event to the view after
a tab was removed from TabControl (somehow). Use notifyDataSetChanged() to be a
better adapter. If we did not call notifyDataSetChanged and getCount() will
return something different, return the old count and post a message to update
the adapter.

Bug: 2307928