Launch intents to open/delete downloads, and handle them in the browser.
Initial work for http://b/issue?id=2384554 : showing all downloads in the
browser.  On the BrowserDownloadPage, no longer rely on the filename,
since once the Browser shows downloads from other applications, the
filenames for those downloads will not be available, and this way
all downloads are handled the same.

Other applications which download files will need to handle the same
Intents as OpenDownloadReceiver in order to open/delete them from
the BrowserDownloadPage.

OpenDownloadReceiver is also a necessary step towards moving the
BrowserDownloadPage into its own application, if we ultimately
decide to do that.