Fix a legit crash that the monkey found.
If the device is running low on memory, the Browser may free background tabs.
removeTab could potentially remove a tab that has a lower index than the current
tab. That would cause mCurrentTab to point to the incorrect tab. Update
mCurrentTab if that is the case so it is pointing to the correct index.

Bug: 2371002