Remove our dependency on gsf.
Hardcode the account values when looking up the hosted domain. We also hardcode
the settings uri for looking up the client id. This is the better solution for
now because the client id is written only once. If the partner setup were to
write the value to the BrowserProvider, it would get erased if the user decided
to reset the Browser.