Use SUGGEST_COLUMN_TEXT_2_URL in browser suggestions
Before, the Browser used an HTML <font> tag to style
URLs in history / bookmark suggestions. The color used
was a hidden framework resource, it assumed that
the caller knew to interpret HTML mark-up, and it only
worked on a light colored background.

Now a separate column is used for displayed URLs, letting
the caller decide how to style URLs.

Part of the fix for http://b/issue?id=2380681

Change-Id: I1c3981de22bccf9cba2f2b903225ef1b50226758