Request icons in the background.
BrowserBookmarksAdapter does not need to request bookmark icons as most of the
bookmarks have up-to-date icons. We still need to request all icons on start up
because a new icon may be loaded as a result of clicking a history items.

Move the icon request to a background thread to not block the ui thread on large
db lookups.

Remove MostVisitedActivity as it is no longer used.

Bug: 2496872
Change-Id: I378dce8fdb7e825da96594cd31b720ea24758af1