Force a load when getting a new intent from voice search.
When reusing an app tab, we generally do not reload in the case
where the URL matches the old URL.  However, in the case of a
voice search, both URLs are "", because the information is held
in UrlData.mVoiceIntent.  If that field is not null, load the
Tab, which will load the new voice search.

Change-Id: I0ba60b1e482cf1fd1b95e4dab969527f18096cc1