Lessen the difference in appearance of title bar during load.
Change to using textfield_search_empty assets, which have margin
values that better match the assets used outside of a load.  As
a result, there is no longer a need to edit the LayoutParams in
TitleBar.java.  However, do set the padding in TitleBar.java, since
it needs to be set for each of the different backgrounds used.

Fixes http://b/issue?id=2558919

Change-Id: Idf0d8bea59bcc76c32e1e03bf736a4b64bf31887