Allow interaction with page while Find is up.
In order to do this, I have changed the FindDialog from
an actual Dialog, which steals all touch events, to
a Linearlayout, which rests below the WebView.

Also dismiss Find when the user opens/closes a subwindow,
or navigates to a new page.

res/layout/browser_subwindow.xml:
Add an id to the holder for the subwindow, so it can be
used to add the FindDialog.

res/values/styles.xml:
Remove the style for FindDialog, as the animations are
now added in code (since FindDialog is now a LinearLayout,
which has no theme).

res/values/themes.xml:
Deleted, as the only theme there was FindDialog, which has
been removed.

src/com/android/browser/BrowserActivity.java:
closeFind is now the starting point for removing the
FindDialog, so that it can be called from Tab.
Close the FindDialog when a new page starts loading.
Call showFind on the current Tab.

src/com/android/browser/FindDialog.java
Change from a Dialog to a LinearLayout, so it can be
inserted into the layout.
Call closeFind directly, which now calls dismiss.
Perform the animations which were previously part
of the theme.
Remove the call to set the height of the find dialog,
which is no longer necessary.
Open and close the IME when opening and closing Find.

src/com/android/browser/Tab.java
Change pointer to mContainer to a LinearLayout, which is
used to add the FindDialog.
Add a pointer to BrowserActivity to SubWindowClient, which
is then used to close the FindDialog in onPageStarted.
Close find when adding/removing a Tab or its subwindow.
Add showFind, which attaches it to the layout for the tab,
and closeFind, which removes it from the layout.

Requires a change to frameworks/base

Change-Id: If6745fb65c5628da827781a7b98061e87b279844