Update patch set 1
Patch Set 1: Verified; Looks good to me, approved

The underlying code was deeply modified in froyo in a way that'll make this change unmergeable, such that it'll only be short-lived. I'm merging it anyway as it'll help manage warnings, but you need to expect it to disappear.

Patch-set: 1
Reviewer: Jean-Baptiste Queru <1000660@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=+2
Label: Verified=+1