Improve the visibility and discoverability of the "add shortcut to home" functionality in the Browser.
When the user selects the bookmark button adjacent to the URL bar
they will be prompted to either add a new bookmark for the current
page or add a shortcut to the current page to their homescreen,
rather than being taken to the bookmark management activity. The
bookmarks button on the options menu will still take the user
directly to the bookmark management activity.

Bug: b/2794945

Change-Id: I07190250379f1d6e2fe6b8ea280317949cd58b15