Difficult to bookmark History items on ldpi-screens
Add a History item as a bookmark by touching the star
is difficult since the touch area of the star is of the
size of the star. The problem is more apparent on
ldpi-screens since the star is small and it's hard to hit
with a finger. The browser loads the url of the History
item if the star is missed.

Enlarging the touch area of the star to cover the whole
right part of the History item solves the problem. The
user doesn't has to hit the star only the right part of
the History item.

Change-Id: I87b8fc555896cf88eac805c96bf31981c2aacbf5