Update patch set 1
Patch Set 1:

problem exists on small screens, where there are two problems. 

The first is that there is no padding between the next / prev button 
and left of the text field. When the display is small the use of the 
linear layout and lack of padding pushes the buttons and text field 
together.

The second problem is that the height of the input field will not be 
the same hight as the buttons if the images are smaller than the 
minimum hight of the input field (since EditText uses wrap_content).

Patch-set: 1