AutoFill Profile editor UI initial checkin
Make a start on a simple AutoFill profile editor so that
we can get rid of the default John Smith profile we've
been using for testing. This CL also moves the autofill
settings from the Privacy Preferences page to the Personal
Preferences page.

Note that this is just the profile editor in the Browser
and although the data entered is persisted to disk, it is
not yet synced with the native AutoFill code so for now we
continue to use the canned profile native side.

Change-Id: Ie609c5f25fffc914267605efdb10444379c29388