Removal of non functional star in context menu of HistoryItem
In contrary to the contextmenus created by longpressing an element
under the "Bookmarks tab" or the "Most visited tab" in the browser,
the context menu created from an element under the "History tab"
has a star present in the header. Removed this star to get a more
unified look and feel. This star also crashed the browser when it
was pressed.

Change-Id: Icddd6ba7d813773445d4dedea589f70be06e5881