Pass webview's useragent when downloading via context menu
Currently null is passed as useragent to the Download Provider when
downloading via Browser context menu (causes the Download
Provider to fallback to the default Download Provider useragent).
This commit simply passes the webview's useragent instead.

This applies to the usecase when choosing save image/link
via the context menu (long press).

Change-Id: Ie9cc5e81630d0dcefe4708980146cc5ed867e4ad