Use picturelistener instead of delay when creating thumbnails
When a download of a bookmarked page has finished the browser
will create a thumbnail of the page. This was done after a
delay of 500 ms to let the Browser have time to create the
picture of the page needed in order to create a thumbnail.
The delay would not guarantee that the browser was ready with
a new picture, which could result in picture of a blank page
or wrong page. Now a picturelistener is instead added when the
download is finished. The picturelistener will receive an
onNewPicture when the picture is available, which will trigger
the creation of a thumbnail.

Change-Id: If615e00221dd034df7bc6598e5694e7cf131491c