More edits to the add bookmark page.
Bug:2953515

BreadCrumbView:
Make sure the height is at least as tall as the separator,
so the height will not change when a separator is added or
removed.

Override getBaseline() to return the baseline of the
crumbs.

AddBookmarkPage:
Move the Add new folder button to the top.

Update the divider between the title and the rest of the
dialog.

Change-Id: I60d329626bd4ec2975cbf6ef00081338beeb7553