Update the "Add to" field of the add bookmarks page.
Bug:2953515

Only change mCurrentFolder when(if) the user presses OK on
a folder.

Allow a cancel press to go back to the default view without
changing mCurrentFolder.

In completeFolderNaming, use the id found in the top bread
crumb to determine the containing folder.

When changing mCurrentFolder, update the "Add to" field.

Also fix a null pointer exception when editing a bookmark.

Change-Id: I6f33b918a88d15e045007f213fa6f39d567ce10b