Update patch set 1
Patch Set 1:

Indeed, the code fragment being modified here was deleted in Google's internal tree (and it's not obvious to me at a glance whether this change is still necessary and where it should be made). This merges and builds fine in the open-source tree.

Patrick, when you feel ready to resolve the merge conflict, please mark this change as Verified, and submit it. Thanks!

Patch-set: 1
Reviewer: Jean-Baptiste Queru <1000660@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=0
Label: Verified=0