use new API (and not db) to access download notifications
instead of opening a cursor on the database to retrieve the file
to be opened when a notification is clicked on, use
DownloadManager's newly introduced API to get Uri and its
mimetype.
also, in BrowserActivity, when downloading a file,
allow it to be MediaScanner scannable and thus allow
media files to be displayed in Gallery.
by default, downloaded files DO NOT appear in Gallery
app any longer

depends on the following 3 CLs
   I1f5dd734e394db0056579a3a0c26862fee27981e
   I5c062ad6d1b58306044cee49ff3827e908d27fd9
   Ia15000de4a66e8728b43fc53f428e098503b003b

Change-Id: Iad11a63fe0a7b8de188d1b6dc0445ccb96211fb2