Update homescreen icon generator to look better on xlarge devices.
Copy mipmap-hdpi resources into mipmap-xlarge to provide larger assets
for xlarge devices (previously we were picking them up from
mipmap-mdpi). With the larger icons, we also need to increase the size
of the rectangle we draw the favicon onto (otherwise we see the points
of the star coming out of the favicon).

Bug: 3224340

Change-Id: If0a297836c288104a32e5dbcd5d886d5849f9698