Facilitate mirroring of the Browser TitleBar
The touch event handling code of the TitleBar
makes assumptions where the button is located
and where the input field is located. This makes
it hard to mirror the layout (implementing BIDI support)
without making changes to the java code.

This change makes it easier to implement mirroring
by checking if the touch event is inside the button
or input field making it independent of where they
are placed inside the TitleBar.

Change-Id: Ic78d7340ecdd16e838c4f169e48e8cb57bd825b0