popular url test should load website list from external file
There are two reasons to use hardcoded external path as website
list: 1) the InstrumentationTestRunner did not provide a
mechanism for passing extra parameters down to test class, so
it's hard to provide test class with a location of test list file
in a parameter 2) if we include the test list file as part of the
asset and loaded via context during test, then it become hard for
test harness to inject a list at test execution time because test
list file is essentially compiled in when the apk is generated.
So I'm reverting the test list loading part back to previous
approach.

Change-Id: I5532de8d153172771dba160d6ad89092f81ec6e4