Making it possible to style browser using overlay
Today it is not possible to style the browser in a good way
using the overlay mechanism. The reason for this is that the
browser is missing an application theme that can be overridden.

The solution is to add a BrowserDefaultTheme that can be
overridden using overlay.

It also adds a default theme to the AddBookmark activity enabling
it to also be styled using the overlay mechaism.

Change-Id: I61bb86e47eb38e55c73ec6e4813e9c864cf3f740