Different approach to auto signin.
On startup, attempt to log the user into google sites.  Show a progress dialog
that the user can cancel if login takes too long.

Add a new preference for toggling auto signin.  This preference shows the
current account and allows the user to choose the account to use.  If there are
no accounts, the option is disabled.  The saved account is validated each time
it is accessed in case the account was removed.

Bug: 3278072
Change-Id: I10ce1dc57a683b2820b17ef6955577037c82f332