Add a "disable autofill" checkbox to the setup prompt
Make the intent of the buttons in the setup autofill dialog
clearer by adding an option to disable the feature. The cancel
button now simply dismisses the dialog and we only disable the
feature if the user checks the new box and clicks OK. Clicking OK
without checking the box takes the user to the prdoile editor as
before.

Bug:3348828
Change-Id: Ib54b879b6953d0c8865d50374188059a43a8e6ba