Clear session cookies before attempting pre-login.
If ClientLogin issues session cookies, we do not want to clear them immediately
after getting them from login.  If we are not going to restore tabs, go ahead
and clear the cookies before attempting pre-login.  Keep track of the tab to
restore so that we don't need to figure it out again.  Requires a change in
frameworks/base that exposes the CookieManager api.

If we receive a 403 from IssueTokenAuth, inval the auth tokens and try again.

Bug: 3421214
Change-Id: I5dd4cc0eba365a20a731ac43dd2571ef6274eaa9