Fix file upload handling after ActivityNotFoundException.
When we launch the file picker for uploads, if nothing can handle
the type of file then we catch the ActivityNotFoundException so
that we do not crash and launch a default picker instead. However
we still will receive a callback for the intent that failed so in
this case we must skip processing it.

Bug:3447924
Change-Id: Ic309f4ba1a6feda981fcce5231b9aed9be70eef2