Add autocomplete to the browser UrlInputView.
Code and tests based on the google search app. A lot
less code would be duplicated if we could somehow
override AutoCompleteTextView but that is made impossible
by it calling a bunch of stuff in its constructor. To do
so would require changes to the existing API.

I've verified that the unit test passes, but other browser
tests appear to fail - even on a clean branch with none
of my changes.

Also fixes a minor bug in SearchEngines.getSearchableInfo( ).

Change-Id: Ic61bc6b8fa27cd210a45dc181ebf15accf503244