Make sure timers are running during pre-login.
Pause/resume timers is really broken. If the user exits the Browser, the process
is still around with paused timers. If pre-login needs to run, timers will be
stalled during the process. This is because there is no current tab during
onResume. Make sure to call resumeTimers and start the cookie sync manager
before handling pre-login.

Bug: 3474608
Change-Id: I12be617947f3ae54a206e31760a2973f81252f50