Cache the factory reset URL.
Hitting the browser provider for the factory reset url is one of
our most common strict mode violations, and triggers sqlite disk
io. So now read it once in a background thread and cache it for
future accesses.

Needed to make the requireInitialization() function static, so we
now wait/notify on BrowserSettings.class rather than the instance.

Bug:5072979

Change-Id: I78550965887e32b4f8ad0eaf0013753e63d6f602