Allow default preload setting to be configured via Gservices.
We now load the default value from a secure setting. To over-ride it, set
Gservices key secure:browser_default_preload_setting to the appropriate
value.

The Gservices default value will only be used if the user has never opened
the 'Bandwidth management' tab in the browswer settings. Once that tab has
been opened, even if the user did not change the setting, the value will be
persisted in the browser and the default will be ignored.

Bug: 5346028

Change-Id: If3b14485dc8bc46e7ed72c49eb5cf3c474893fe3