Return history/bookmarks from BrowserProvider.
This makes the global-search query to the browser return the history
table joined with the bookmarks table. This way, bookmarks in the
history will have the correct icon set.

Note: Since we're doing a left join of history with bookmarks, this
means we won't show bookmarks which haven't been visited, something
we consider a feature. 8)

Bug: 5370935
Change-Id: I0068432ca1a8ef78f27b7a84eb7be3be56eae644