Remove dump V8 counters from browser debug menu
Removing the function from the WebView. This is a private function,
and to do anything useful WebCore had to be recompiled with the
special ANDROID_INSTRUMENTATION option.

Needed by https://android-git.corp.google.com/g/#/c/148912/

There is a bug for fixing ANDROID_INSTRUMENTATION:
bug 2816327

Since this has not been used for more than a year, removing
this option instead of fixing.

Change-Id: Ib5033a377a388d2ab211407aefe4b4932b441b76