UserAgent not propagated from Browser to DownloadProvider
Browser does not propagate the userAgent to DownloadProvider
for downloads. When downloads are started the default userAgent
in DownloadProvider is used instead of using the userAgent of
the Browser. This makes downloads on some services / web pages
fail.

This is solved this by including the Browser userAgent in the
request to the DownloadManager.

Change-Id: If4e0f3c29eb7fd5fc585dd01987ae7f0b54ffc32