Update patch set 1
Patch Set 1:

Selim: True. The snapshot provider has no dependencies to the Browser so it should be able to move that into a separate package (similar to the Download Provider). Do you think it is worth the effort to secure/lockdown the Browser? If so, we could do the move, but you have to judge on how big the delta is between AOSP/master and your internal tree, and if it is possible. Some proposal on how to proceed here would be appreciated.

Patch-set: 1